adjusted sidebar to fit the needs
